setup: Install fonts only on desktops

+++ setup.sh
@@ -74,12 +74,14 @@ if [ -n "$USRSHELL" ] && [ "$USRSHELL" != "zsh" ]; then
 fi
 
 # install JetBrainsMono Nerd Font
-[ ! -d "$HOME/.local/share/fonts" ] && mkdir -p "$HOME/.local/share/fonts"
-if [ ! -f "$HOME/.local/share/fonts/JetBrainsMonoNerdFontMono-Regular.ttf" ]; then
-    curl -LO https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/JetBrainsMono.zip
-    unzip -o -j JetBrainsMono.zip "*.ttf" -d "$HOME/.local/share/fonts"
-    rm JetBrainsMono.zip
-    fc-cache -f
+if [ "$is_desktop" = 1 ]; then
+    [ ! -d "$HOME/.local/share/fonts" ] && mkdir -p "$HOME/.local/share/fonts"
+    if [ ! -f "$HOME/.local/share/fonts/JetBrainsMonoNerdFontMono-Regular.ttf" ]; then
+        curl -LO https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/JetBrainsMono.zip
+        unzip -o -j JetBrainsMono.zip "*.ttf" -d "$HOME/.local/share/fonts"
+        rm JetBrainsMono.zip
+        fc-cache -f
+    fi
 fi
 
 # install github packages
